fix: use default service-linked role for autoscaling
The IAM role `keycloak-ecs-autoscale-role` was not being used by
AWS, as it defaults to using the built-in service-linked role for
the autoscaling namespace. This created persistent drift in
Terraform, which would flag the custom role as a pending change
on plan, but that change would be ignored by AWS on apply.
Since the role is unused, it can safely be deleted.